requirements.txt
gunicorn==20.1.0
dj-database-url==0.5.0
whitenoise==5.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
xlwt==1.3.0
Django==3.2.6
django-heroku==0.3.1
django-crispy-forms==1.12.0
anytree==2.8.0
Pillow>=8.3.1
django-widget-tweaks>=1.4.8
django-cleanup>=5.2.0
cloudinary>=1.17.0
django-cloudinary-storage>=0.2.3
